added classes for pharmacy map view

+++ app/src/main/java/com/example/loginapp/Entity/Pharmacy.java
@@ -6,7 +6,7 @@ public class Pharmacy extends MedicalLocation{
     private String pharmacy_name;
     private String pharmacy_ID;
 
-    public Pharmacy(long Latitude, long Longitude, String pharmacy_ID, String pharmacy_name, String pharmacy_address) {
+    public Pharmacy(double Latitude, double Longitude, String pharmacy_ID, String pharmacy_name, String pharmacy_address) {
         super(Latitude, Longitude);
         this.pharmacy_address=pharmacy_address;
         this.pharmacy_name =pharmacy_name;
